Refactor AdvancedExternalizerAdapter to new wildfly-clustering-marshalling-infinispan module.
