+++ feature-pack/src/main/resources/modules/system/layers/base/org/wildfly/clustering/marshalling/infinispan/main/module.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
+
 <!--
   ~ JBoss, Home of Professional Open Source.
   ~ Copyright 2010, Red Hat, Inc., and individual contributors
@@ -21,25 +22,17 @@
   ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
   -->
 
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
+<module xmlns="urn:jboss:module:1.5" name="org.wildfly.clustering.marshalling.infinispan">
+    <properties>
+        <property name="jboss.api" value="private"/>
+    </properties>
 
-    <parent>
-        <groupId>org.wildfly</groupId>
-        <artifactId>wildfly-clustering</artifactId>
-        <version>11.0.0.Alpha1-SNAPSHOT</version>
-    </parent>
+    <resources>
+        <artifact name="${org.wildfly:wildfly-clustering-marshalling-infinispan}"/>
+    </resources>
 
-    <artifactId>wildfly-clustering-marshalling</artifactId>
-    <packaging>pom</packaging>
-
-    <name>WildFly: Clustering marshalling modules</name>
-
-    <modules>
-        <module>api</module>
-        <module>infinispan</module>
-        <module>jboss</module>
-    </modules>
-</project>
+    <dependencies>
+        <module name="org.wildfly.clustering.marshalling.api"/>
+        <module name="org.infinispan.commons"/>
+    </dependencies>
+</module>
